v2. refactoring, added readme
