+++ .gitignore
@@ -1,3 +1,4 @@
 # Project exclude paths
 /out/
 .idea
+java-warmup-course.iml
