+++ src/lesson_3/Connector.java
@@ -1,4 +1,4 @@
-package lessonThree;
+package lesson_3;
 
 public record Connector(
         String host,
