+++ src/lesson_3/ConnectorMapper.java
@@ -1,4 +1,4 @@
-package lessonThree;
+package lesson_3;
 
 import java.util.List;
 import java.util.stream.Collectors;
